+++ packages/manager/src/features/Images/utils.test.tsx
@@ -2,7 +2,23 @@ import { linodeFactory } from '@linode/utilities';
 
 import { eventFactory, imageFactory } from 'src/factories';
 
-import { getEventsForImages, getImageLabelForLinode } from './utils';
+import {
+  getEventsForImages,
+  getImageLabelForLinode,
+  isLinodeKubeImageId,
+} from './utils';
+
+describe('isLinodeKubeImageId', () => {
+  it('should be false if the image id does not start with linode/', () => {
+    expect(isLinodeKubeImageId('private/15943292')).toBe(false);
+  });
+  it('should be false if the image is a regular non-kube image', () => {
+    expect(isLinodeKubeImageId('linode/alpine3.15')).toBe(false);
+  });
+  it('should be true if the image is a linode kube image', () => {
+    expect(isLinodeKubeImageId('linode/ubuntu24.04-kube-v1.23.4')).toBe(true);
+  });
+});
 
 describe('getImageLabelForLinode', () => {
   it('handles finding an image and getting the label', () => {
